Fix warnings related to CPLError calls
git-svn-id: https://svn.osgeo.org/gdal/trunk@14049 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/gtiff/geotiff.cpp
@@ -3016,7 +3016,7 @@ GDALDataset *GTiffDataset::OpenDir( const char *pszCompositeName )
             {
                 XTIFFClose( hTIFF );
                 CPLError( CE_Failure, CPLE_OpenFailed, 
-                          "Requested directory %d not found.", nOffset );
+                          "Requested directory %lu not found.", (long unsigned int)nOffset );
                 return NULL;
             }
             nOffset--;
